implemented delete expense feature in the app with authroization.

+++ frontend/src/components/Expenses/ExpenseItem.js
@@ -4,12 +4,32 @@ import ExpenseDate from "./ExpenseDate";
 import Card from "../UI/Card";
 import "./ExpenseItem.css";
 import ExpenseContext from "../../store/expense-context";
+import axios from "axios";
+import { API_URL } from "../../utils/var";
 
 const ExpenseItem = (props) => {
   const expenseCtx = useContext(ExpenseContext);
 
-  const deleteHandler = () => {
+  const deleteHandler = async () => {
+    const confirmDelete = window.confirm("Do You Want To Delete This Item?");
+try{
+
+  if (!confirmDelete) return;
+  const response = await axios.delete(
+    `${API_URL}/api/v1/expenses/delete/${props.id}`
+    );
+    console.log(props.id);
+
+    console.log("response.data", response.data);
+    console.log("response.status", response.status);
     expenseCtx.onDelete(props.id);
+  }catch(err){
+    // console.log(err.response.status)
+    if(err.response.status === 401){
+      return alert(err.response.data.message || "Not Authenticated")
+    }
+    alert("Something went wrong")
+  }
   };
 
   return (
